requirements.txt
flask
mysql-connector-python
scikit-learn
